requirements.txt
fastapi~=0.116.1
pydantic~=2.11.7
starlette~=0.47.2
sse-starlette~=2.2.1
pydantic-settings~=2.10.1
dotenv~=0.9.9
python-dotenv~=1.1.1
langchain-openai~=0.3.31
langchain-core~=0.3.74
langchain~=0.3.27
uvicorn
requests~=2.32.5
firebase_admin~=7.1.0
aiohttp~=3.12.15
openai~=1.101.0
numpy~=2.3.2
PyYAML~=6.0.2
torch>=2.9.0
sentence-transformers~=5.1.0
scikit-learn~=1.7.1
transformers~=4.55.4
alembic==1.16.5
psycopg2~=2.9.10
chromadb~=1.0.20
pytest~=8.4.2
pytest-asyncio
pandas~=2.3.2
concurrent_log_handler
googlemaps~=4.10.0
langchain-community~=0.3.27
tinytag~=2.1.2
asyncpg
geoalchemy2~=0.18.0
shapely~=2.1.2
firebase_admin~=7.1.0
python-multipart
pymorphy3
ruwordnet
Pillow~=11.0.0
emoji>=2.0.0
sqlalchemy==1.4.50
